+++ src/platforms/esp/int.h
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <stdint.h>
+#include <stddef.h>
 
 namespace fl {
-    // ESP32: Like ARM platforms, uint32_t and int32_t are 'unsigned long' and 'long'
+    // ESP platforms (ESP8266/ESP32): Like ARM platforms, uint32_t and int32_t are 'unsigned long' and 'long'
     // This matches the Xtensa toolchain behavior similar to ARM toolchains
     typedef short i16;
     typedef unsigned short u16;
@@ -10,8 +12,7 @@ namespace fl {
     typedef unsigned long u32;
     typedef long long i64;
     typedef unsigned long long u64;
-    // size_t is unsigned int on ESP32 (32-bit)
-    typedef unsigned int sz;
-    // uintptr_t is unsigned int on ESP32 (32-bit pointers)
-    typedef unsigned int uptr;
+    // Use standard types directly to ensure exact compatibility
+    typedef size_t sz;
+    typedef uintptr_t uptr;
 }
